pyproject.toml
[tool.poetry]
name = "techdoc-experiment"
version = "0.1.0"
description = "A modernized framework for ingesting local and remote documentation into Port."
authors = ["EricFernandezPort <ericf@getport.io>"]
readme = "README.md"
packages = [{include = "port_tools", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12"
requests = ">=2.32.4,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
markdown = ">=3.8.2,<4.0.0"
beautifulsoup4 = ">=4.13.4,<5.0.0"
python-frontmatter = ">=1.1.0,<2.0.0"
pygithub = ">=2.6.1,<3.0.0"
python-gitlab = ">=6.1.0,<7.0.0"
azure-devops = ">=7.1.0b1"
toml = ">=0.10.2,<0.11.0"
fuzzywuzzy = ">=0.18.0,<0.19.0"
python-levenshtein = ">=0.27.1,<0.28.0"
openai = ">=1.93.0,<2.0.0"
python-dateutil = ">=2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

